# Fix sparse token selection in deepseek v3.2