The following fixes/changes are included in this PR

- Added missing ReLU activation (`mx.maximum(scores, 0)`) in indexer after computing attention scores, as implemented by reference [kernel.py:241](https://github.com/deepseek-ai/DeepSeek-V3.2-Exp/blob/main/inference/kernel.py#L241) to filter out negative attention scores before token selection.

- Changed `scaled_dot_product_attention(..., cache=cache)` to `cache=cache[0]` to pass `KVCache` object instead of `CacheList` wrapper

- Added `scores.sum(axis=1)` in indexer to aggregate per-head scores into unified token selection (matching [kernel.py:244](https://github.com/deepseek-ai/DeepSeek-V3.2-Exp/blob/main/inference/kernel.py#L244)), changing output shape from `(B, heads, L, topk)` to `(B, L, topk)` and fixing repetitive generation caused by each head selecting different tokens.

- Rewrote attention layer's sparse token handling to create boolean masks instead of per-head direct indexing, since indexer now returns unified indices for all heads rather than per-head indices.

- Unified decode/prefill sparse masking paths to handle chunked prefill correctly by always using `k_seq = keys.shape[2]` instead of assuming square masks.