@kernelpool 2025-10-07
Added some supporting benchmarks and evaluations using **DeepSeek-V3.2-Exp-3bit**

# Benchmarks

`mlx_lm.benchmark --model /Volumes/WD_EXTRA/models/catalyst/DeepSeek-V3.2-Exp-3bit -p <n> -g 128`

  Prompt Processing Speed (tokens/sec)

  | Context Length | Without PR | With PR | Improvement |
  |----------------|------------|---------|-------------|
  | 512 tokens     | 173.030    | 172.763 | -0.15%      |
  | 1024 tokens    | 199.939    | 199.922 | -0.01%      |
  | 2048 tokens    | 209.701    | 209.912 | +0.10%      |
  | 4096 tokens    | 157.607    | 187.479 | +18.9% ✅    |

  Generation Speed (tokens/sec)

  | Context Length | Without PR | With PR | Improvement |
  |----------------|------------|---------|-------------|
  | 512 tokens     | 21.157     | 21.162  | +0.02%      |
  | 1024 tokens    | 19.675     | 19.565  | -0.56%      |
  | 2048 tokens    | 7.668      | 15.005  | +95.7% ✅    |
  | 4096 tokens    | 7.628      | 12.131  | +59.0% ✅    |

  Peak Memory Usage (GB)

  | Context Length | Without PR | With PR | Improvement |
  |----------------|------------|---------|-------------|
  | 512 tokens     | 298.082    | 298.093 | 0.00%       |
  | 1024 tokens    | 301.330    | 301.330 | 0.00%       |
  | 2048 tokens    | 308.636    | 308.636 | 0.00%       |
  | 4096 tokens    | 327.799    | 318.847 | -8.95 GB ✅  |
  
# Perplexity
  
  `python ppl.py --model /Volumes/WD_EXTRA/models/catalyst/DeepSeek-V3.2-Exp-3bit --sequence-length <n>`

  | Seq Length | Without PR | With PR | Delta    | Time (s) | Peak Mem (GB) | Tokens/sec | Samples |
  |------------|------------|---------|----------|----------|---------------|------------|---------|
  | 512        | 4.477      | 4.477   | 0.000    | 436      | 279           | 250        | 213     |
  | 1024       | 3.847      | 3.847   | 0.000    | 436      | 279           | 500        | 213     |
  | 2048       | 3.589      | 3.589   | 0.000    | 437      | 279           | 1000       | 213     |
  | 4096       | ⚠️ 725.967 | 3.872   | -722.095 | 439      | 279           | 1991       | 213     |