@awni 2025-10-07
> I'm very surprised by the speedup you are observing in token generation speed from this. I'd like to understand that better.

I think removing the take on the keys and values and using the boolean mask is what sped this up.

Very nice though. Would you mind addressing the comments and then we can merge it?